src/test/java/com/betting/bets/stake/impl/ScoreStakeTest.java
package com.betting.bets.stake.impl;

import com.betting.bets.stake.Stake;
import com.betting.bets.stake.StakeOutcome;
import com.betting.bets.stake.StakeTest;
import com.betting.results.EventResults;
import com.betting.results.ResultPair;
import com.betting.results.type.ResultType;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;
import org.mockito.Mock;
import org.springframework.boot.test.context.SpringBootTest;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.when;

@SpringBootTest
class ScoreStakeTest extends StakeTest {
    private Stake stake;
    @Mock
    private EventResults eventResults;
    static Stream<Arguments> winArgumentsProvider() {
        return Stream.of(
                Arguments.of("1:1", ResultPair.of(1,1)),
                Arguments.of("2:1", ResultPair.of(2,1)),
                Arguments.of("1:2", ResultPair.of(1,2))
        );
    }
    static Stream<Arguments> loseArgumentsProvider() {
        return Stream.of(
                Arguments.of("1:1", ResultPair.of(1,0)),
                Arguments.of("2:1", ResultPair.of(2,2)),
                Arguments.of("1:2", ResultPair.of(1,1))
        );
    }

    @ParameterizedTest
    @MethodSource("winArgumentsProvider")
    void testResolveOutcomeWin(String stakeName, ResultPair score) {
        stake = new ScoreStake(stakeName, stakeType, ResultType.ROUNDS);
        when(eventResults.getOpponentsPoints()).thenReturn(score);
        stake.resolveOutcome(eventResults);
        assertEquals(StakeOutcome.WIN, stake.getStakeOutcome());
    }
    @ParameterizedTest
    @MethodSource("loseArgumentsProvider")
    void testResolveOutcomeLose(String stakeName, ResultPair score) {
        stake = new ScoreStake(stakeName, stakeType, ResultType.ROUNDS);
        when(eventResults.getOpponentsPoints()).thenReturn(score);
        stake.resolveOutcome(eventResults);
        assertEquals(StakeOutcome.LOSE, stake.getStakeOutcome());
    }
}